+++ server/src/lib/events.js
@@ -6,7 +6,6 @@ const eventEmitter = new EventEmitter();
 events.on = eventEmitter.on.bind(eventEmitter);
 events.once = eventEmitter.once.bind(eventEmitter);
 events.emit = (eventName, eventPayload) => {
-  console.log('hello world');
   eventEmitter.emit(eventName, eventPayload);
 };
 events.names = eventEmitter.eventNames.bind(eventEmitter);
